VoiceMap/backend/requirements.txt
# Dependencies for the VoiceMap backend
flask
openai
requests
python-dotenv
werkzeug
